#26 - check if file exists before trying to move them

+++ src/new-installer/mce-install-common.sh
@@ -583,8 +583,9 @@ function RemoveOfflineSource {
 function MangleStartupFiles {
 
 	if [[ "$c_startupType" == "1" ]] ;then
-		mv /etc/init.d/gdm /etc/init.d/gdm.saved
-		mv /etc/init.d/kdm /etc/init.d/kdm.saved
+		# Make sure non of the other display manager starts
+		[ -f /etc/init.d/gdm ] && mv /etc/init.d/gdm /etc/init.d/gdm.saved
+		[ -f /etc/init.d/kdm ]mv /etc/init.d/kdm /etc/init.d/kdm.saved
 		echo '#!/bin/bash
 	if [[ "$1" == start ]]; then
 		/usr/pluto/bin/Startup_Core-Hybrid.sh
